stepper: allow specifying endstop_of on rails
When using rails, it can be beneficial to specify the endstop for a
given motor, such that it follow another motor on the rail.

To allow the general case to work, where endstops from both higher and
lower indexed rail steppers can be used on any other rail stepper, a
two-pass approach is used. All steppers are added, and steppers with
endstop_pointers are then linked to the correct endstops.

Signed-off-by: Lasse Dalegaard <dalegaard@gmail.com>